Move userinfo to _app.tsx
